[update] rename decode_inst_(16|32) to inst_(16|32)
